+++ src/math/CVector.cpp
@@ -7,10 +7,10 @@
    modify, merge, publish, distribute, sublicense, and/or sell copies
    of the Software, and to permit persons to whom the Software is
    furnished to do so, subject to the following conditions:
-   
+
    The above copyright notice and this permission notice shall be
    included in all copies or substantial portions of the Software.
-   
+
    THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
    EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
    MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -46,7 +46,7 @@ namespace CGF{
     streams  = new cudaStream_t[n_devices];
     partial_reductions = 0;
 
-    for(uint i=0;i<n_devices;i++){
+    for(int i=0;i<n_devices;i++){
       d_data[i]   = 0;
       d_mapped[i] = 0;
       d_mapped_reductions[i] = 0;
@@ -81,7 +81,7 @@ namespace CGF{
     streams  = new cudaStream_t[n_devices];
     partial_reductions = 0;
 
-    for(uint i=0;i<n_devices;i++){
+    for(int i=0;i<n_devices;i++){
       d_data[i]   = 0;
       d_mapped[i] = 0;
       d_mapped_reductions[i] = 0;
@@ -106,7 +106,7 @@ namespace CGF{
   }
 
   template<class T>
-  CVector<T>::CVector(ulong s, const ThreadPool* p, bool c):
+  CVector<T>::CVector(int s, const ThreadPool* p, bool c):
     CObject(p), vec(0), copy(c){
 
     d_data   = new T*[n_devices];
@@ -114,15 +114,15 @@ namespace CGF{
     d_mapped_reductions = new T*[n_devices];
     streams  = new cudaStream_t[n_devices];
     partial_reductions = 0;
-    
-    for(uint i=0;i<n_devices;i++){
+
+    for(int i=0;i<n_devices;i++){
       d_data[i]   = 0;
       d_mapped[i] = 0;
       d_mapped_reductions[i] = 0;
     }
 
     pinned_memory = 0;
-    
+
     if(s%16==0){
       size = s;
     }else{
@@ -135,7 +135,7 @@ namespace CGF{
 
     if(pool == 0){
       allocateDevice(0);
-    }    
+    }
 
     if(copy && pool){
       /*Allocate host proxy vector*/
@@ -151,7 +151,7 @@ namespace CGF{
     }
     if(copy && pool){
       if(n_devices != 1){
-	cudaSafeCall(cudaFreeHost(pinned_memory));
+        cudaSafeCall(cudaFreeHost(pinned_memory));
       }
     }
     if(partial_reductions){
@@ -180,10 +180,10 @@ namespace CGF{
     cgfassert(origSize == vec->getSize());
 
     /*Copy data*/
-    for(uint i=0;i<n_devices;i++){
+    for(int i=0;i<n_devices;i++){
       cudaSafeCall(cudaMemcpy(d_data[i], vec->data + vRange[i].startBlock,
-			      sizeof(T)*vRange[i].range, 
-			      cudaMemcpyHostToDevice));      
+                              sizeof(T)*(uint)vRange[i].range,
+                              cudaMemcpyHostToDevice));
     }
   }
 
@@ -196,21 +196,21 @@ namespace CGF{
   template<class T>
   void CVector<T>::allocateReductionBuffers(){
 #if 0
-    int flags = 
+    int flags =
       cudaHostAllocDefault;
     /*      cudaHostAllocMapped |
-      cudaHostAllocWriteCombined |
-      cudaHostAllocPortable;*/
-    cudaSafeCall(cudaHostAlloc((void**)&partial_reductions, 
-    			       sizeof(T)*n_devices, 
-			       flags));
+            cudaHostAllocWriteCombined |
+            cudaHostAllocPortable;*/
+    cudaSafeCall(cudaHostAlloc((void**)&partial_reductions,
+                               sizeof(T)*n_devices,
+                               flags));
 #else
     partial_reductions = new T[n_devices];
 #endif
 
     if(pool == 0){
       /*      cudaSafeCall(cudaHostGetDevicePointer((void**)&d_mapped_reductions[0],
-	      (void*) &(partial_reductions[0]),0));*/
+              (void*) &(partial_reductions[0]),0));*/
     }
   }
 
@@ -220,93 +220,93 @@ namespace CGF{
       return;
 
 #ifdef PINNED
-    int flags = 
+    uint flags =
       cudaHostAllocMapped |
       cudaHostAllocWriteCombined |
       cudaHostAllocPortable;
-        
-    cudaSafeCall(cudaHostAlloc((void**)&pinned_memory, 
-			       sizeof(T)*size, 
-			       flags));
+
+    cudaSafeCall(cudaHostAlloc((void**)&pinned_memory,
+                               sizeof(T)*(uint)size,
+                               flags));
 #else
-    cudaSafeCall(cudaMallocHost((void**)&pinned_memory, 
-				sizeof(T)*size));
+    cudaSafeCall(cudaMallocHost((void**)&pinned_memory,
+                                sizeof(T)*(uint)size));
 #endif
   }
 
   template<class T>
   void CVector<T>::setDefaultRange(){
     /*Each device get the same ammount of data*/
-    uint segments = size/16;
-    uint segmentsPerDevice = segments/n_devices;
+    int segments = size/16;
+    int segmentsPerDevice = segments/n_devices;
 
-    uint total_segments = 0;
-    for(uint i=0;i<n_devices;i++){
+    int total_segments = 0;
+    for(int i=0;i<n_devices;i++){
       if(i==n_devices-1){
-	segmentsPerDevice += segments%n_devices;
+        segmentsPerDevice += segments%n_devices;
       }
       vRange[i].range      = segmentsPerDevice * 16;
       vRange[i].startBlock = total_segments    * 16;
       vRange[i].endBlock   = total_segments    * 16 + segmentsPerDevice * 16;
-      
+
       mRange[i].range    = segmentsPerDevice   * 16;
       mRange[i].startRow = total_segments      * 16;
       mRange[i].endRow   = total_segments      * 16 + segmentsPerDevice * 16;
-      
+
       total_segments += segmentsPerDevice;
     }
   }
 
   template<class T>
   void CVector<T>::allocateDevice(const Thread* caller){
-    uint tid = 0;
+    int tid = 0;
     if(caller){
       tid = caller->getId();
     }
 
     if(copy && pool){
       /*Each device has its own full copy of the vector*/
-      uint lsize = 0;
-      for(uint i=0;i<n_devices;i++){
-	lsize += vRange[i].range;
+      int lsize = 0;
+      for(int i=0;i<n_devices;i++){
+        lsize += vRange[i].range;
       }
 
       lsize = size; /*Added for square matrices*/
 
       textureSize = lsize;
 
-      cudaSafeMalloc((void**)&(d_data[tid]), 
-		     sizeof(T)*lsize);
+      cudaSafeMalloc((void**)&(d_data[tid]),
+                     sizeof(T)*(uint)lsize);
 
       if(vec != 0){
-	cudaSafeCall(cudaMemcpy(d_data[tid], vec->data,
-				sizeof(T)*size, 
-				cudaMemcpyHostToDevice));
+        cudaSafeCall(cudaMemcpy(d_data[tid], vec->data,
+                                sizeof(T)*(uint)size,
+                                cudaMemcpyHostToDevice));
       }
 
       if(n_devices != 1){
 #ifdef PINNED
-	/*Create mapped device pointers*/
-	cudaSafeCall(cudaHostGetDevicePointer((void**)&d_mapped[tid],
-					      (void*)  pinned_memory,0));
+        /*Create mapped device pointers*/
+        cudaSafeCall(cudaHostGetDevicePointer((void**)&d_mapped[tid],
+                                              (void*)  pinned_memory,0));
 #endif
       }
     }else{
       /*Each device has only a part of the complete vector*/
-      cudaSafeMalloc((void**)&(d_data[tid]), 
-		     sizeof(T)*vRange[tid].range);
+      cudaSafeMalloc((void**)&(d_data[tid]),
+                     sizeof(T)*(uint)vRange[tid].range);
       if(vec != 0){
-	cudaSafeCall(cudaMemcpy(d_data[tid], vec->data + vRange[tid].startBlock,
-				sizeof(T)*vRange[tid].range, 
-				cudaMemcpyHostToDevice));
+        cudaSafeCall(cudaMemcpy(d_data[tid], vec->data + vRange[tid].startBlock,
+                                sizeof(T)*(uint)vRange[tid].range,
+                                cudaMemcpyHostToDevice));
       }
 
       textureSize = vRange[tid].range;
     }
-    
+
     if(partial_reductions){
       /*cudaSafeCall(cudaHostGetDevicePointer((void**)&d_mapped_reductions[tid],
-	(void*) &(partial_reductions[tid]),0));*/
+        (void*) &(partial_reductions[tid]),0));*/
     }
 
     cudaStreamCreate(&streams[tid]);
@@ -314,7 +314,7 @@ namespace CGF{
 
   template<class T>
   void CVector<T>::deallocateDevice(const Thread* caller){
-    uint tid = 0;
+    int tid = 0;
     if(caller)
       tid = caller->getId();
 
@@ -329,27 +329,27 @@ namespace CGF{
 
   template<class T>
   void CVector<T>::print(const Thread* caller){
-    uint tid = 0;
+    int tid = 0;
     if(caller)
       tid = caller->getId();
 
     if(copy){
       if(tid == 0){
-	cudaDisplayFloatArray<T>(d_data[tid], size,size);
+        cudaDisplayFloatArray<T>(d_data[tid], size, size);
       }
       getchar();
       if(tid == 1){
-	cudaDisplayFloatArray<T>(d_data[tid], size,size);
+        cudaDisplayFloatArray<T>(d_data[tid], size, size);
       }
     }else{
-      uint sz = vRange[tid].range;
-      
+      int sz = vRange[tid].range;
+
       if(tid == 0){
-	cudaDisplayFloatArray<T>(d_data[tid], sz,sz);
+        cudaDisplayFloatArray<T>(d_data[tid], sz, sz);
       }
       getchar();
       if(tid == 1){
-	cudaDisplayFloatArray<T>(d_data[tid], sz,sz);
+        cudaDisplayFloatArray<T>(d_data[tid], sz, sz);
       }
       getchar();
     }
@@ -359,35 +359,35 @@ namespace CGF{
   T CVector<T>::sum(const Thread* caller)const{
     cgfassert(this != reductionBuffer);
     cgfassert(reductionBuffer != 0);
-    
+
     T* result;
     if(pool == 0){
-      parallel_reduction<T>(this, reductionBuffer, &result, 
-			    /*d_mapped_reductions[0]*/0, caller);
-      
-      cudaSafeCall(cudaMemcpy(&(partial_reductions[0]), 
-			      reductionBuffer->d_data[0],
-			      sizeof(T), 
-			      cudaMemcpyDeviceToHost));
+      parallel_reduction<T>(this, reductionBuffer, &result,
+                            /*d_mapped_reductions[0]*/0, caller);
+
+      cudaSafeCall(cudaMemcpy(&(partial_reductions[0]),
+                              reductionBuffer->d_data[0],
+                              sizeof(T),
+                              cudaMemcpyDeviceToHost));
 
       return partial_reductions[0];
     }else{
       cgfassert(caller != 0);
-      
-      uint tid = caller->getId();
-      parallel_reduction<T>(this, reductionBuffer, &result, 
-			    /*d_mapped_reductions[tid]*/0, caller);
-      
-      cudaSafeCall(cudaMemcpy(&(partial_reductions[tid]), 
-			      /*reductionBuffer->d_data[tid]*/result,
-			      sizeof(T),
-			      cudaMemcpyDeviceToHost));
+
+      int tid = caller->getId();
+      parallel_reduction<T>(this, reductionBuffer, &result,
+                            /*d_mapped_reductions[tid]*/0, caller);
+
+      cudaSafeCall(cudaMemcpy(&(partial_reductions[tid]),
+                              /*reductionBuffer->d_data[tid]*/result,
+                              sizeof(T),
+                              cudaMemcpyDeviceToHost));
 
       caller->sync();
-      
+
       T total_sum = 0;
-      for(uint i=0;i<n_devices;i++){
-	total_sum += partial_reductions[i];
+      for(int i=0;i<n_devices;i++){
+        total_sum += partial_reductions[i];
       }
       return total_sum;
     }
@@ -402,9 +402,9 @@ namespace CGF{
 
     /*Copy vector associated with this thread to pinned_memory*/
     cudaSafeCall(cudaMemcpyAsync(pinned_memory + vRange[TID].startBlock,
-				 d_data[TID] + vRange[TID].startBlock,
-				 sizeof(T)*vRange[TID].range,
-				 cudaMemcpyDeviceToHost, streams[TID]));
+                                 d_data[TID] + vRange[TID].startBlock,
+                                 sizeof(T)*(uint)vRange[TID].range,
+                                 cudaMemcpyDeviceToHost, streams[TID]));
 #ifdef BENCHMARK
     cudaThreadSynchronize();
 #endif
@@ -417,15 +417,15 @@ namespace CGF{
     if(n_devices == 1 || copy == false)
       return;
 
-    for(uint i=0;i<n_devices;i++){
+    for(int i=0;i<n_devices;i++){
       if(i==TID){
-	continue;
+        continue;
       }
       /*Update vectors on other devices*/
       cudaSafeCall(cudaMemcpyAsync(d_data[TID] + vRange[i].startBlock,
-				   pinned_memory + vRange[i].startBlock,
-				   sizeof(T)*vRange[i].range,
-				   cudaMemcpyHostToDevice, streams[TID]));
+                                   pinned_memory + vRange[i].startBlock,
+                                   sizeof(T)*(uint)vRange[i].range,
+                                   cudaMemcpyHostToDevice, streams[TID]));
     }
     /*In order to measure the elapsed time correctly*/
 #ifdef BENCHMARK
